move RunningStepBase to structure_handler.py
